Replicate code from sushi and modify per food
Code taken from sushi/rewritten to the personalized needs of each food,
which is french toast and scrambled eggs. They didn't have it before,
thus it was added.

Another function added was an anchor tag with an href that led back to
the homepage from each recipe because there was a problem of mobility.